@@ -2,43 +2,43 @@
     "name": "laravel/laravel",
     "type": "project",
     "description": "The Laravel Framework.",
-    "keywords": ["framework", "laravel"],
+    "keywords": [
+        "framework",
+        "laravel"
+    ],
     "license": "MIT",
     "require": {
         "php": "^8.3.0",
         "ext-pdo": "*",
         "awcodes/filament-gravatar": "^2.0",
         "bezhansalleh/filament-shield": "^3.1",
-        "croustibat/filament-jobs-monitor": "^2.1",
-        "doctrine/dbal": "^3.5",
-        "filament/filament": "^3.1",
-        "guzzlehttp/guzzle": "^7.2",
-        "honeybadger-io/honeybadger-laravel": "^3.15",
+        "croustibat/filament-jobs-monitor": "^2.3",
+        "filament/filament": "^3.2",
+        "honeybadger-io/honeybadger-laravel": "^4.0",
         "josiasmontag/laravel-recaptchav3": "dev-master",
-        "laravel/framework": "^10.0",
-        "laravel/sanctum": "^3.2",
-        "laravel/tinker": "^2.8",
+        "laravel/framework": "^11.20",
+        "laravel/sanctum": "^4.0",
+        "laravel/tinker": "^2.9",
         "league/flysystem-aws-s3-v3": "^3.0",
         "rickdbcn/filament-email": "^1.0",
         "shuvroroy/filament-spatie-laravel-backup": "^2.0",
         "shuvroroy/filament-spatie-laravel-health": "^2.0",
         "spatie/cpu-load-health-check": "^1.0",
-        "spatie/laravel-backup": "^8.1",
+        "spatie/laravel-backup": "^8.6",
         "spatie/laravel-permission": "^6.2",
         "spatie/security-advisories-health-check": "^1.1",
         "stechstudio/filament-impersonate": "^3.5",
-        "usernotnull/tall-toasts": "^2.0"
+        "usernotnull/tall-toasts": "^2.1"
     },
     "require-dev": {
-        "barryvdh/laravel-ide-helper": "^2.13",
-        "fakerphp/faker": "^1.21",
+        "barryvdh/laravel-ide-helper": "^3.0",
+        "fakerphp/faker": "^1.23",
         "laravel/pint": "^1.13",
-        "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^7.0",
-        "pestphp/pest": "^2.0",
-        "pestphp/pest-plugin-laravel": "^2.0",
-        "pestphp/pest-plugin-livewire": "^2.1",
-        "spatie/laravel-ignition": "^2.0"
+        "mockery/mockery": "^1.6",
+        "nunomaduro/collision": "^8.0",
+        "pestphp/pest": "^2.35",
+        "pestphp/pest-plugin-laravel": "^2.2",
+        "pestphp/pest-plugin-livewire": "^2.1"
     },
     "autoload": {
         "psr-4": {
