@@ -11,8 +11,11 @@ class AMPDSCode extends Model
     protected $table = 'ampds_codes';
     protected $fillable = ['code', 'description', 'arrest_code', 'far_code'];
 
-    protected $casts = [
-        'arrest_code' => 'boolean',
-        'far_code' => 'boolean',
-    ];
+    protected function casts(): array
+    {
+        return [
+            'arrest_code' => 'boolean',
+            'far_code' => 'boolean',
+        ];
+    }
 }
